grid修复有符号无符号数混用

+++ frameworks/core/components_ng/pattern/grid/grid_scroll/grid_scroll_layout_algorithm.cpp
@@ -262,7 +262,7 @@ void GridScrollLayoutAlgorithm::InitialItemsCrossSize(
 
     if (crossCount_ != crossLens.size()) {
         crossCount_ = crossLens.size();
-        gridLayoutInfo_.crossCount_ = crossCount_;
+        gridLayoutInfo_.crossCount_ = static_cast<int32_t>(crossCount_);
     }
 
     int32_t index = 0;
@@ -468,7 +468,7 @@ void GridScrollLayoutAlgorithm::FillBlankAtEnd(
                 --currentIndex;
                 break;
             }
-            i += childState - 1;
+            i += static_cast<uint32_t>(childState) - 1;
             // Step3. Measure [GridItem]
             LargeItemLineHeight(itemWrapper, hasNormalItem);
             gridLayoutInfo_.endIndex_ = currentIndex;
@@ -987,11 +987,11 @@ void GridScrollLayoutAlgorithm::SkipRegularLines(bool forward)
         return;
     }
     int32_t estimatedLines = gridLayoutInfo_.currentOffset_ / lineHeight;
-    if (forward && gridLayoutInfo_.startIndex_ < (estimatedLines) * crossCount_) {
+    if (forward && gridLayoutInfo_.startIndex_ < estimatedLines * static_cast<int32_t>(crossCount_)) {
         gridLayoutInfo_.startIndex_ = 0;
         gridLayoutInfo_.currentOffset_ = 0;
     } else {
-        gridLayoutInfo_.startIndex_ -= estimatedLines * crossCount_;
+        gridLayoutInfo_.startIndex_ -= estimatedLines * static_cast<int32_t>(crossCount_);
         gridLayoutInfo_.currentOffset_ -= lineHeight * estimatedLines;
     }
 }
